simulation & test_world: move coordinate system to middle of plate + various other changes
The coordinate system starts in the middle of the plate. That allows to
sum the force vectors to get the total pressure put on the plate.

Furthermore the last commit introduced some errors that could have been
found at compile time - e.g. also adding the const qualifier to the
implementation of that function. Those were fixed too