voter: inital commit with nanomsg
nanomsg was choosen because it claims to provide several communication
patterns -- e.g. a publish/suscribe protocol, which is a nice fit for
the world <-> controller communication.

Furthermore it is a plain C library, so the students don't silently get
a dependency on C++ runtime libraries.

Supposedly there is a .NET port, I'll have to check whether that works.

An RPC framework would be nice (and there are some hobby projects on
github) but if the number of functions stays this low, I can probably do
without

nanomsg also has a netcat like application that can be used for simple
tests:
$ nanocat --rep --bind tcp://127.0.0.1:8000 --data pong --raw
$ nanocat --pub --bind tcp://127.0.0.1:8001 --data server -i 5