controller: use algorithm to send movement request
This should be the second to last change as far as the network is concerned.
Right now unity sends all data to the voter. The voter sends it to the
controllers.

The controllers use the algorithm to generate movement requests and then
send off one. That is recvd by the voter.

Right now the voter still needs to parse it and do the actual voting.