Reworked the "path finding" algo. And all its parts
I really can't think of anything wrong in this part. Though it
leads to very weird results, e.g. as the robot turns in the correct
direction (as evidenced by a visible raycast in the simulation) the
angle to the target point also increases. And then for some reason
the robot suddenly switches to the other direction and then switches
back and ...

It just never reaches the goal